requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bitarray==3.4.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
dnspython==2.7.0
email-validator==2.2.0
env==0.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
jinja2==3.1.6
langcodes==3.5.0
language-data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.0
packaging==25.0
pdfreader==0.1.15
pillow==11.2.1
preshed==3.0.10
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pymongo==4.13.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
safetensors==0.5.3
scalar-fastapi==1.0.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-alignments==0.9.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.49.0
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
